pyproject.toml
[project]
name = "rag-chatbot-git"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
